+++ source/test-suite/fixtures/TupleConstructorExpected.hs
@@ -1,0 +1,1 @@
+pair = (,) <$> Just 1 <*> Just 2